+++ lib/src/config_resolver.rs
@@ -433,7 +433,6 @@ mod tests {
     use indoc::indoc;
 
     use super::*;
-    use crate::config::ConfigSource;
 
     #[test]
     fn test_expand_home() {
@@ -1006,10 +1005,10 @@ mod tests {
             hostname: "",
         };
         let resolved_config = resolve(&source_config, &context).unwrap();
-        insta::assert_snapshot!(resolved_config.layers()[0].data, @r#"
+        insta::assert_snapshot!(resolved_config.layers()[0].data, @r"
         a = 'a none'
         b = 'b none'
-        "#);
+        ");
         if cfg!(target_os = "linux") {
             assert_eq!(resolved_config.layers().len(), 3);
             insta::assert_snapshot!(resolved_config.layers()[1].data, @"a = 'a linux'");